rename(apps/app): 공통 유틸 함수 common 폴더로 이동
